requirements.txt
# Core
ccxt>=4.1.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0
py_vollib>=1.0.1
arch>=6.2.0
statsmodels>=0.14.0

# Storage
clickhouse-connect>=0.7.0
redis>=5.0.0

# API & Dashboard
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
plotly>=5.17.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
